API V1.1

+++ app/controllers/comics_controller.rb
@@ -2,6 +2,19 @@ class ComicsController < ApplicationController
     protect_from_forgery with: :null_session
     before_action :authorize_request
 
+    def index
+        if params[:user_id]
+            comics = User.find(params[:user_id]).comics
+            display_comics = comics.map do |comic| 
+                    return {id: comic.id,
+                            title: comic.title,
+                            description: comic.description,
+                            thumbnail_url: comic.thumbnail_url }
+                end
+            render json: {comics: display_comics}, status: 200 
+        end
+    end
+
     def create 
         new_comic = @current_user.comics.build(comic_params)
         if new_comic.save
@@ -13,10 +26,12 @@ class ComicsController < ApplicationController
 
     def update 
         comic = Comic.find(params[:id])
-        if comic.update(comic_params)
-            render json: {message: 'Successfully updated'}, status: 200 
-        else
-            render json: {error: comic.errors}, status: 422
+        if comic_params[:title] || comic_params[:description]
+            if comic.update(comic_params)
+                render json: {message: 'Successfully updated'}, status: 200 
+            else
+                render json: {error: comic.errors}, status: 422
+            end
         end
     end
 
@@ -36,9 +51,11 @@ class ComicsController < ApplicationController
 
     def destroy 
         comic = Comic.find(params[:id])
-        if comic
+        if comic && current_user === comic.user
             comic.destroy
             render json: { message: "Successfully destroyed" }, status: 200 
+        elsif current_user.id != comic.user_id
+            render json: { error: {comic:["can't be deleted by anyone but its author"]} }, status: 422 
         else
             render json: { error: {comic:["not found"]} }, status: 422 
         end
@@ -47,6 +64,6 @@ class ComicsController < ApplicationController
     private
 
     def comic_params 
-        params.require(:comic).permit(:title, :description)
+        params.require(:comic).permit(:title, :description, :thumbnail)
     end
 end
